Rework job processing in light of Rails 5's Reloader, see #3221
