+++ lib/sidekiq/version.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 module Sidekiq
-  VERSION = "4.2.5"
+  VERSION = "4.3.0"
 end
